feat: adiciona configuracao swr
