+++ apps/admin/app/layout.tsx
@@ -1,14 +1,20 @@
+"use client";
+
+import { api } from "@/lib/api";
 import "@mantine/carousel/styles.css";
 import { ColorSchemeScript, MantineProvider } from "@mantine/core";
 import "@mantine/core/styles.css";
 import "@mantine/notifications/styles.css";
-import { ClientOnlyCookiesProvider } from "@repo/components";
 import { theme } from "@repo/constants";
+import { CookiesProvider } from "react-cookie";
+import { SWRConfig } from "swr";
 
-export const metadata = {
-  title: "BenefitEaseAdmin",
-  description: "Administração dos benefícios de forma simples e eficiente.",
+const defaultSetOptions = {
+  path: "/",
+  secure: process.env.NODE_ENV === "production",
 };
+
+const fetcher = (url: string) => api.get(url).then((res) => res.data);
 
 export default function RootLayout({
   children,
@@ -26,7 +32,11 @@ export default function RootLayout({
       </head>
       <body>
         <MantineProvider defaultColorScheme="auto" theme={theme}>
-          <ClientOnlyCookiesProvider>{children}</ClientOnlyCookiesProvider>
+          <CookiesProvider defaultSetOptions={defaultSetOptions}>
+            <SWRConfig value={{ fetcher, errorRetryCount: 3 }}>
+              {children}
+            </SWRConfig>
+          </CookiesProvider>
         </MantineProvider>
       </body>
     </html>
