+++ packages/components/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./client-only-cookies-provider";
 export * from "./header";
 export * from "./with-auth";
 export * from "./without-auth";
